+++ Sources/ReaderThumbsView.h
@@ -1,9 +1,9 @@
 //
 //	ReaderThumbsView.h
-//	Reader v2.6.0
+//	Reader v2.7.1
 //
 //	Created by Julius Oklamcak on 2011-09-01.
-//	Copyright © 2011-2012 Julius Oklamcak. All rights reserved.
+//	Copyright © 2011-2013 Julius Oklamcak. All rights reserved.
 //
 //	Permission is hereby granted, free of charge, to any person obtaining a copy
 //	of this software and associated documentation files (the "Software"), to deal
@@ -51,7 +51,7 @@
 
 @interface ReaderThumbsView : UIScrollView
 
-@property (nonatomic, unsafe_unretained, readwrite) id <ReaderThumbsViewDelegate> delegate;
+@property (nonatomic, weak, readwrite) id <ReaderThumbsViewDelegate> delegate;
 
 - (void)setThumbSize:(CGSize)thumbSize;